+++ public_html/console/migrations/m170920_072725_alter_users_and_ads_table.php
@@ -7,7 +7,7 @@ class m170920_072725_alter_users_and_ads_table extends Migration
     public function up()
     {
         $this->addColumn('users', 'phone_number', $this->string()->null());
-        $this->addColumn('ads', 'expiry_date', $this->integer()->notNull());
+        $this->addColumn('ads', 'expiry_date', $this->bigInteger()->notNull());
         $this->addColumn('ads', 'url', $this->string()->notNull()->unique());
         $this->addColumn('ads', 'placements_id', $this->integer(10)->unsigned()->notNull());
